+++ lib/index.d.ts
@@ -1,3 +1,5 @@
 export { CharInfoEx } from "./CharInfoEx";
 export { CharInfoOffset } from "./CharInfoOffset";
 export { getStringUTF16NullTerminated } from "./Helpers";
+export { ColorIndex, MiiColor } from "./ColorIndex";
+export { ColorOrderLTD, ColorOrderMiiStudio } from "./ColorOrder";