+++ CHANGELOG.md
@@ -1,8 +1,20 @@
 # Changelog
 
+## [0.0.4] - 2026-04-10
+
+- Add the color array and color order data.
+- Add turning CharInfoEx back into a ArrayBuffer
+
+
+## [0.0.3] - 2026-04-09
+
+- Add `fromJson` to return `CharInfoEx | CharInfoEx[]`.
+
+
 ## [0.0.2] - 2026-04-09
 
 - Fix array loop having wrong value
+
 
 ## [0.0.1] - 2026-04-09